rpm: Fix variable names
The package list for rpmtree "foo" is supposed to be defined
by a variable called "$foo_base". In a few cases we are not
doing that, so fix the inconsistencies.

This leads to a few awkward variable names, specifically when
the name of the rpmtree ends in "base". We're going to
improve things in the next commit.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>