node-labeller.sh: Consider AppArmor restrictions
Even though the virt-handler pod is privileged, on the systems with
AppArmor there might be a host profile which will be automatically
picked for the /usr/sbin/libvirtd binary. That may block the execution
of /usr/libexec/qemu-kvm. In such a case, try moving the qemu executable
to a location, which is more common for AppArmor-enabled Linux distros.

Signed-off-by: Vasiliy Ulyanov <vulyanov@suse.de>