rpm: Add virtio-gpu-pci device for aarch64
This device was part of the qemu-kvm-core package in CentOS
Stream 8, but in CentOS Stream 9 additional effort has been put
into modularizing QEMU and as a consequence it now lives in its
own package, which we need to bring in explicitly in order to
avoid regressions.

Note that the device is going to be available on aarch64 only:
x86_64 VMs could technically use the device for secondary
displays, but there is currently no way to configure a KubeVirt
VM to use non-default video devices, and for x86_64 the default
is VGA.

Similarly, the non-PCI version of the device is omitted because
KubeVirt will always use a PCI topology for aarch64 VMs, and
the OpenGL-accelerated version is left out because there is no
way to turn on acceleration in the context of KubeVirt.

Reported-by: Howard Zhang <howard.zhang@arm.com>
Signed-off-by: Andrea Bolognani <abologna@redhat.com>