client-go: Update client to use namespaced expand-vm-spec
The kubevirt client-go is updated to use the namespaced expand-vm-spec
endpoint and a unit test is added for verification.

Direct use of the ExpandSpec struct is replaced with the
ExpandSpecInterface to be consistent with other subresource endpoints.

Conflicts:
  staging/src/kubevirt.io/client-go/kubecli/instancetype_test.go

NOTE(0xFelix): The conflict is due to not having the api proxy feature
in release-0.58. Therefore the tests for proxied URLs are removed.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>