guest-agent, tests: Increase timeout fetching pod IPs
When testing pod IPs reported by the guest-agent on the VMI
the tests wait for a default timeout to get the updated IPs.
Since the status update depends on guest agent polling and propagation
between virt-launcher and virt-handler which can take more time
the Expect timeout should allow more time in order to make sure
that the IPs are picked up by the test.
Timeout is set to 2 minutes which is the guest agent steady state
polling time

Signed-off-by: Ram Lavi <ralavi@redhat.com>