virt-launcher: Assure the domain is controlled by the same VMI
Introduce a validation mechanism to assure that the VMI which triggered
the domain creation is the only one who can continue to reconcile it.

On creation, the domain UUID is set based on the VMI Firmware UUID
(vmi.Spec.Domain.Firmware.UUID) and on any future access
to it (through SyncVMI) it is validated to match.

The VMI Firmware UUID is setting the domain SMBIOS sysinfo-system-uuid,
which the domain will also use as its general UUID.

Signed-off-by: Edward Haas <edwardh@redhat.com>