Add to snapshot status lists of excluded/included volumes in the snapshot
This will be used to help the user understand if a volume he expected
to have in the snapshot was not a part of the snapshot.
The snapshot is desgined to be ready to use once all the volumes
that can be snapshoted and have appropriate snapshot storage classs
are snapshoted.
The user for some reason might miss the fact that it has volumes which
are on a not snapshotable storage class.
The user will be able to see that those volumes were not a part of the
snapshot in this list.

Signed-off-by: Shelly Kagan <skagan@redhat.com>