Solve flakiness for guestfs tests
Sometimes, the libguestfs pod is deleted by the test clean-up but the
function runGuestfsOnPVC hasn't gracefully terminated yet. The function
runGuestfsOnPVC has a goroutine that returns when the channel of the
fakeCreateAttacher is closed.

With the new refactoring, a new channel has been added to select between
an error in the guestfs routine and to gracefully terminates the test.

Signed-off-by: Alice Frosi <afrosi@redhat.com>