Avoid empty dir for the appliance directory
With the previous method, the appliance was extracted at runtime, and
located in `/usr/local/lib/guestfs/appliance` path. This directory was
backed by an emptyDir. This method isn't working anymore because it
overwrites the extracted appliance with an empty directory. Therefore,
we remove the use of the empty dir for the appliance.
Instead, we use it for `HOME` were libguestfs needs the permissions to
create and write files.

Signed-off-by: Alice Frosi <afrosi@redhat.com>