instancetype: Add support for preference resource requirements
This change introduces basic support for expressing minimum resource
requirements for CPU and Memory within preferences.

These requirements are checked against the VirtualMachine or an
associated Instance type by the validation webhook on admission.

For CPU resources the preferredCPUTopology is taken into account when
checking a VirtualMachine without an Instance type. The default
`PreferSockets` behaviour being used if no preference is defined. As
such a CPU resource requirement of 1 vCPU within the preference will
result in a check for CPU topology containing at least 1 Socket within
the VirtualMachine.

Co-authored-by: Felix Matouschek <fmatouschek@redhat.com>
Signed-off-by: Lee Yarwood <lyarwood@redhat.com>