Clean unschedulable nodes
Currently CleanNodes() does not restore the nodes if a test only changes
the `spec.unschedulable` field. This can happen for example in a migration
test, which set all control-plane nodes unschedulable, in a cluster where
all nodes are both control-plane and worker. This results in a test failure,
and since the nodes are not properly restored, the subsequent tests fail on
cascade.

Signed-off-by: fossedihelm <ffossemo@redhat.com>