virt-api: Add test for empty namespace in expand_test.go
This adds a unit test which sets an empty namespace as path parameter
for a call to the ExpandSpecRequestHandler. This is a theoretical
situation which should not be technically possible at the moment.

Setting of the namespace path parameter in the following test was
removed as it was redundant. callExpandSpecApi already does it.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>