virt-launcher: Introduce the metadata cache
The metadata package implements a cache with which data can be stored
and loaded in a safe concurrent manner.

The metadata cache is introduced to replace the existing
libvirt-metadata store.
The following change will gradually transition the metadata to the new
cache.

Decoupling from the libvirt-metadata store has several benefits:
- Eliminates the need to access the domain in order to read and save
metadata information. It is simpler and faster.
- Removes the direct dependency on libvirt.

Signed-off-by: Edward Haas <edwardh@redhat.com>