Fix: don't expect node to be always stopped
In HyperV tests in the "TSC frequency is not exposed on the cluster"
case, at the BeforeEach(), we stop node labeller on a node if TSC is
exposed to fake as if it's not exposed.

In the AfterEach(), we then expect node-labeller to be stopped,
although it's not always the case (as said above, if TSC is not
exposed in the first place, we don't stop node-labeller).

Signed-off-by: Itamar Holder <iholder@redhat.com>