instancetype: Look for default annotations on Data{Volume,Source}s
This change tweaks the behaviour of
inferDefault{Instancetype,Preference} so that we now look directly for
any default annotations on DataVolumes or DataSources referenced by the
Volume identified by InferFromVolume. When these annotations are not
found the previous behaviour of looking for an underlying PVC with
annotations is used.

This should allow operators and users to directly decorate the higher
level DataVolume and DataSource resources when available without having
to also decorate the underlying PVCs that they might not create
directly.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>