Add VM definition URL to export server so you can get complete VM information from export.
The url path is defined in the status of the VMExport object but
in general the format is https://<service>.<namespace>.svc/export-def

This will return yaml containing two or three things:
1. The VM definition yaml which is completely rendered without using
instance types or preferences.
2. A config map containing the CA cert of the http endpoint (internal)
or the CA cert of the Route/Ingress (external)
3. (Optional) Any datavolumes associated with the VM that are not part
of the original VM definition DataVolumeTemplates. This also includes
persistent volumes not associated with the templates. These DataVolumes
allow the user to download the images with CDI.

For the DataVolumes (and DataVolumeTemplates) the source will be set to
http, and the URL will point to the appropriate endpoint to download the
image. The endpoint will be the gzipped version of the disk image to
improve transfer time.

The /export-def endpoint is behind the same token as the volumes themselves
and should be provided as a query paramter or a header just like when
getting the disk images.

One additional https endpoint is available to get a secret that is compatible
with CDI. This gives CDI all the information need to import from the VMExport
into whatever target cluster CDI is running in.

The https endpoint can take a parameter called 'externalURI' which indicates
the returned yaml should point to the external link disk image instead of
the internal one. Any value other than blank ("") will result in the links
pointing to the external urls. If you connect through the virtexport-proxy,
the proxy will automatically add this parameter so the user doesn't have to
do so.

Signed-off-by: Alexander Wels <awels@redhat.com>