virt-launcher: Remove references to the libvirt metadata
All special treatment to libvirt metadata is removed.
No metadata is persisted into libvirt and no read of it is done.

In addition, on migration, all metadata is dropped (not only the
migration subset).

Special care was required for data that previously has been treated
by the conversion step (translation between VMI to Domain spec).
Specifically, the UID and the grace-period data is now handled outside
the main converter step.

In order to completely remove all traces of the libvirt metadata, the
metadata structure needs to be removed from the `Domain.Spec` and placed
elsewhere.
This change is to be followed up, as it involves also the virt-handler,
which has backward compatibility challenges.

Signed-off-by: Edward Haas <edwardh@redhat.com>