Safely consume TLSSecurityProfile from APIServer CR (#2149)
APIServer CR is validating TLSSecurityProfile
less strictly than CNAO.
HCO in general delegates the validation
to the operators that is managing,
but this is not true for values directly
read from the APIServer CR
that are not passing into our validating
webhook.
So if the cluster admin requires
a custom configuration for TLSSecurityProfile
on the APIServer CR all the components
but CNAO will accept it while CNAO
will refuse it continuously.
Let's prevent this sanitizing the input.
this is not validated

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=2137896

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>