tests, storage/restore: Remove a redundant prompt login
A redundant VM console prompt login has been removed.
It caused some flakes in the tests, when the second login was attempted.

Although for the Fedora guest, detecting a login should have been
supported, it seems that from time to time the detection is not working
as expected. Reducing double logins may help reduce the flakes a bit.

Signed-off-by: Edward Haas <edwardh@redhat.com>