e2e tests: virt-handler canary upgrade test improvment
Recently the test was observerd with high flakiness. The suggested improvement
is prone to events that can repeat a few times i.e. update event can happen multiple times
with maxUnavailable:1 or maxUnavailable:10% because not only the expected fields are being updated
during the rollout. Another improvement is the sequence of events - it may happen that the rollout
will finish earlier than the restore of maxUnavailable back to 1,
for example if there is some API latency between virt-operator and API server,
and there are few worker nodes, it may happen that more than one pod will be updated
under the maxUnavailable:1 configuration.

Signed-off-by: enp0s3 <ibezukh@redhat.com>