Correctly set defaults for golang clients (#2151)
Correctly set defaults for generated golang clients
using also '// +default' comment marker.
Please notice that +default and +kubebuilder:default
syntaxes are slightly different.
See:
https://book.kubebuilder.io/reference/markers.html#marker-syntax
and
https://github.com/kubernetes/kube-openapi/blob/master/test/integration/testdata/defaults/default.go

Fixing the usage of pointers/direct values in the golang
API remebering that:
- pointer allows distinguishing unset from the zero value for that type
- structs are not omitted from encoder output even where omitempty is specified

Exception to this general rules are required on pointers
to typedef aliases beacuse generated defaults lose type
information for aliases.
See:
https://github.com/kubernetes/gengo/issues/235

Avoid duplicating and hardcording defaults in
pkg/components/components.go since now we can properly
consume them at golang level.

Amend a few tests to properly consume the expected defaults.

This is not going to break backward compatibility at CRD level,
while it will break golang one but, on the other side,
we know that golang clients were not able to properly
consume defaults in a clean way so this is definitively
an enhancement.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>