Tolerate DV being missing warning
Due to datavolume garbage collection, there's a small race where we may
inspect a PVC which refers to a DV, but the DV is already gone, causing
spurious and short-lived errors.

We've opted not to fix this race in this PR:
https://github.com/kubevirt/containerized-data-importer/pull/2487

Tolerate the warning showing up - this is a rare occurrence.

Signed-off-by: Maya Rashish <mrashish@redhat.com>