vm-killer-image, Add nonroot user to container image
Currently, in order to run jobs with specific non-root users, the
container spec is set with the userID.
Such a configuration is used on the vm-killer-image image.
This however requires higher SCC priviliges for the jobs running
this image.
In order to avoid this:
- Configuring the image to use a non root user on the container
dockerfile image level (created by bazel).
- Removing the RunAsUser field from the container manifest in
renderContainerSpec().

Signed-off-by: Ram Lavi <ralavi@redhat.com>