Add tls to the vsock websocket connection
If clients set `tls=true` on the websocket call to connect to a vsock
service, virt-handler will wrap the vsock connection with a TLS
connection and will try to authenticate with the current virt-handler
certificate.

Agents in the guest OS can query the system service on the connection
attempt for the current CA and verify the ceritifcate againts it.

This basically allows guest agents to ensure that a connection is coming
from a service which has at least the CAP_NET_BIND_SERVICE privilege,
since only with that privilege the system service which provides the CA
bundle can be established.

Signed-off-by: Roman Mohr <rmohr@google.com>