instancetype: Ignore unexpected existing CRs during restore
5cc62c0f8f1fc368ffc920ab29fda7b55272728f failed to accommodate repeat
attempts to reconcile a VirtualMachineRestore that in turn lead to
multiple calls to restoreInstancetypeControllerRevision being made for
the same ControllerRevision.

This change handles this case by ignoring any existing
ControllerRevisions found during the restore, assuming that the existing
ControllerRevision was created by a previous call to
restoreInstancetypeControllerRevision.

Future changes will check the contents of this ControllerRevision
against that of the VirtualMachineSnapshot to ensure they match.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>