network plumbing: Use non-ordinal naming for virt-launcher pod interfaces
As for today, virt-launcher pod interface names derived from the order
of the interfaces in the VMI spec.

In order to enable unplug interfaces functionality, the pod interfaces
name scheme must change, as if we remove an interface from the VMI spec
it makes it imposable to associate between the VMI spec interfaces and
the pod interface names.

Use a different, non-ordinal based, algorithm to generate the pod
interface names (requested from Multus).

Up to now we requested these interfaces named `net1`, ..., `netX`,
depending on the order of the given interface within the VMI spec.

This commit breaks that pattern, using non-ordinal naming (compute
the hash of the  VMI spec network name, pruned to 11 characters).

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>
Signed-off-by: Or Mergi <ormergi@redhat.com>