network: rename pod iface names
Use a different - non ordinal based - algorithm to generate the pod
interface names requested from multus.

Up to now we requested these interfaces named `net1`, ..., `netX`,
depending on the ordinal of the given interface within the VMI spec.

This commit breaks that pattern, using something non-ordinal (compute
the hash of the interface name, pruned to 11 characters).

Virt-handler needs to perform an extra step for non-root VMIs with
macvtap interfaces: it needs to set the owner of the device files as the
pod user. Thus, to identify the ifindex of a pod interface, it was
needed to update the algorithm to find the corresponding pod interface.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>