virt-handler: Replace CGO linking with a Go alternative
It is preferable to avoid CGO linking in virt-handler and limit it as
much as possible to the virt-launcher (where it must link with the
libvirt lib).

This is sometimes impossible, but in this case, the stable Linux API can
be partially duplicated into the native Go code.

The CGO linkage had side effects on packages which imported the
`virthandler` package, causing build issues.
(specifically, it collided with the e2e test binary)

Signed-off-by: Edward Haas <edwardh@redhat.com>