tests, storage, export: Fix VM start helper
The VM start helper is using ginkgo `Eventually` to retry the spec
update. In the retry flow, the vm variable is updated even when the
update fails, which in practice initializes it.

Update the retry loop by storing the VM namespace and name upfront and
using the stored names to fetch a fresh object content.

Signed-off-by: Edward Haas <edwardh@redhat.com>