Reject vm if there is vmi with the same name
VMI and VM are closely related resources.
The name of the VMI created from a VM is the same of the VM itself.
This might cause some inconsistencies when you create a VM and
there is already a VMI with the same name. This results in the association
between the two resources, but actually they are not.
This commit "implements" the association of the two, checking the existence
of a VMI with the same name, during the VM creation.

Signed-off-by: fossedihelm <ffossemo@redhat.com>