Reject vmi if there is vm with the same name
VMI and VM are closely related resources.
The name of the VMI created from a VM is the same of the VM itself.
This might cause some inconsistencies when you create a VMI and
there is already a VM with the same name. This results in the association
between the two resources, but actually they are not.
This commit "implements" the association of the two, checking the existence
of a VM with the same name, during the VMI creation.

Signed-off-by: fossedihelm <ffossemo@redhat.com>