Wait for the DV to be successful before we inspect the VMI
If kubevirt beats CDI's pod to using the PVC, it will try to
create the disk.img itself with the host-disk code.

kubevirt's host-disk code doesn't behave nicely when local
storage is used: it assumes that it can use the full capacity
seen in the PVC status capacity.

Unfortunately, in the case of local storage PVs, the capacity
is shared among all the PVs, and it isn't possible to use the
full size.

This causes the vmi to appear initially as Failed, but perhaps
if we waited for CDI to finish importing, the VMI will start.

Signed-off-by: Maya Rashish <mrashish@redhat.com>