pointer: Introduce a generic pointer package
The pointer package has a single P function that returns the address of
the given value.

It is aimed to replace `k8s.io/utils/pointer` with a generic version.
Unlike the previous version, the generic one accepts any type.

To illustrate its usage, the initial usage is done in the e2e tests, at
the libvmi package.

Signed-off-by: Edward Haas <edwardh@redhat.com>