Add foreground finalizer to virtual machine
Currently, when deleting a vm which has an associated vmi,
the vm object is immediately removed, while the vmi object wait
for the pod deletion first.
This can cause a misunderstanding in case the vmi is not deleted
due to some errors. For example, the subsequent creation of a vm
with the same name results in the adoption of the pending vmi.
Adding a finalizer on the vm object prevents new vm creation with
the same name to be avoided, since the original vm object remains
in a terminating state, until the vmi object is gone.

Increasing the removeNamespace timeout.

Signed-off-by: fossedihelm <ffossemo@redhat.com>