virt-launcher: Add debug logs when the cache metadata is updated
When libvirt was used to store the metadata information, debugging could
depend on libvirt logging or its direct access (with virsh).

Now that this is no longer available, V(4) logs are added to allow
debugging if the need arises.

The default V(2) is used for the migration data in case of failure.

Signed-off-by: Edward Haas <edwardh@redhat.com>