macvtap: fix ownership change of the tap interface
Virt-handler needs to perform an extra step for non-root VMIs with
macvtap interfaces: it needs to set the owner of the device files as the
pod user. Thus, to identify the ifindex of a pod interface, it was
needed to update the algorithm to find the corresponding pod interface.

Before this patch we were attempting to change the ownership of the file
pointed to by the name of the multus network - i.e. the network
attachment definition name. This was wrong: we need to use the name of
the pod interface (macvlan side). This only worked in CI since the name
of the net-attach-def used is `net1` - which matches the name of the pod
interface [0], since the algorithm requests for multus interfaces named
`net<X>` where X is the ordinal of the interface in the
`vmi.Spec.Domain.Devices.Interfaces` slice.

[0] - https://github.com/kubevirt/kubevirt/blob/c2a498a00f4f9ecc974f547db7db24f5a5f728bd/tests/network/macvtap.go#L66

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>