Fix podman detection
It is possible to create a Podman socket that understands
Docker API (see https://stackoverflow.com/a/72693093). This
is valuable, for example, when trying to use Podman inside
Podman/Docker.

Currently, to detect if Podman exists, we check if the
podman.sock exists. Since in most scenarios such a socket
doesn't exist, this logic doesn't make sense.

Now, to detect if Podman is available, we execute "podman ps"
and test its exit code - in the exact same way we detect Docker
existance.

Signed-off-by: Itamar Holder <iholder@redhat.com>