Close HTTP body and fully read it on POST requests
This is necesarry for connection re-use. As
client.Do()'s documentation says:

"If the returned error is nil, the Response will contain
a non-nil Body which the user is expected to close. If
the Body is not both read to EOF and closed, the
Client's underlying RoundTripper (typically Transport)
may not be able to re-use a persistent TCP connection
to the server for a subsequent "keep-alive" request."

Source: https://pkg.go.dev/net/http#Client.Do

Signed-off-by: Itamar Holder <iholder@redhat.com>