Append workload scheduling constraints to the virt-launcher pod
Append cluster-workload-pod affinity and cluster-workload-preferred-node
affinity to the virt-launcher pods when configured on the KubeVirt CR.

Before to that these fields were just ignored, since they were only
applied to virt-handler, but virt-handler is a daemonset and there these
values have no meaning.

Now all scheduling constraints which have no meaning for virt-handler
are now appended to the virt-launcher pod for both, migration, as well
as starting VMs.

Signed-off-by: Roman Mohr <rmohr@google.com>