Improve the distroless bump script
To make the script idempotent, we need to work around the fact that
buildozer returns a non-zero exit code (3), if the commands succeeded,
but the file was not changed. Now, first the digest is replaced with a
timestamp, to force buildozer in having to change the file, even if the
actual digest stayed the same since the last run.

Further, since that leads to more buildozer invocations, the commands
are changed to take multiple instructions from stdin, to do batch
modifications.

Signed-off-by: Roman Mohr <rmohr@google.com>