Fix the "when reconciling ovs-cni" unit test (#2191)
The "when reconciling ovs-cni" unit test in
controllers/operators/networkAddons_test.go is wrong. It runs a
scenario, but check the result against the wrong object - the original
object before the change, instead of the object the it reads back from
the client.

This bug hides a wrong test scenario, where the OVS field is expected to
be found but it's not.

This PR fixes the test to check the NetworkAddons CR as read from the
client, and the wrong test scenario.

This PR also changes `Expect(cl.Get(...)).ToNot(HaveOccurred())` to
`Expect(cl.Get(...)).To(Succeed())`, for better readability.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>