Refactor: Add NewManagerFromPath() to cgroup package
Before this commit, it's only possible to create
a cgroup manager for a VMI (or more accurately,
for the virt-launcher's compute container's cgroup).

NewManagerFromPath() allows to initialize a cgroup
manager from a cgroup path. With this function it's
possible to create a manager arbitrarily from any
cgroup of any container.

In addition, a new consts.go file is introduced.

Signed-off-by: Itamar Holder <iholder@redhat.com>