virt-launcher: Cleanup explicit socket definition
Currently, in session mode we are explicitly telling which socket
to use. The same thing does not happen, however, in system mode,
in which we leave to libvirt itself the burden of choosing the socket.
We can homologate the behaviors, removing the explicit socket in session mode.

Also removed an unused parameter in `startDomainEventMonitoring` function.

Signed-off-by: fossedihelm <ffossemo@redhat.com>