ksd: Support passing BaseDomain on Openshift (#2198)
* Update auto gen files

It updates only the year.
Result of make sanity

Signed-off-by: Or Shoval <oshoval@redhat.com>

* ksd: Support passing BaseDomain on Openshift

When running OpenShift the cluster name is available
via cluster DNS config resource.
Fetch it and pass it to KSD if it is deployed.
KSD will use it in order to build the FQDN.
This way different clusters will have different FQDN
suffix.

Signed-off-by: Or Shoval <oshoval@redhat.com>

Signed-off-by: Or Shoval <oshoval@redhat.com>