network, setup: Separate pod network state machine logic
During the pod network configuration, a state machine is used to assure
that once the network have been modified, no future changes are allowed.

This change separates the state logic from the other details (e.g. how
to configure the network, what data is collected and how it is stored
for future use).

The separation will allow future changes to be better isolated (e.g. how
to configure the pod network) from the rest of the steps.

Signed-off-by: Edward Haas <edwardh@redhat.com>