Do not inject LimitRange defaults into VMI
Currently when LimitRange is configured with defaults for compute resources such as CPU or memory,
these defaults are being injected to the VMI spec early in the VMI mutating stage.
Hence, when looking at the VMI spec we can't really know the source of information
for these values, whether they are user-defined or not.

This situation causes issues at the stage of launcher-pod rendering. The resource renderer assumes
the VMI spec values are user defined and prefers them over caclulated resource requests.

If the user defines CPU topology with CPU overcommit, the calculated CPU request is being ignored
in favor to the CPU request from the LimitRange.

For example: a VMI with configured CPU topology of 4 vCPU in spec.domain.CPU.cores
will have a resource request of 400m (num_of_cores * 1000 / AllocationRatio).
Lets take LimitRange Default CPU request as 100m.
The renderred virt-launcher pod will have 100m CPU request instead of 400m.

The proposed solution is to ignore the LimitRange defaults, giving priority
to user configuration, and later to calculated configuration.

Signed-off-by: enp0s3 <ibezukh@redhat.com>