network, setup: Use options to pass the handler to VMNetworkConfigurator
To assist testing, a special factory was use to pass a customized
handler when creating the `VMNetworkConfigurator` object.

In order to allow future customizations to the object, this change
replaces the factory with options (a build pattern).

From a code aesthetics point of view, this pattern allows the production
code to avoid the use of special factories or options
(by using defaults) while the testing code to add them and
override these defaults.

Signed-off-by: Edward Haas <edwardh@redhat.com>