network: Introduce an independent procsys (net) driver
The driver is aimed to provide an adapter layer between kubevirt and
the proc/sys (sysfs) backend implementation.

A fake is also included, allowing testing upper logic.

The content is mainly copied from `driver/common.go`.
The original is due to be removed once the setup flow stops using it.

Signed-off-by: Edward Haas <edwardh@redhat.com>