network: Introduce an independent virt-chroot driver
The driver is aimed to provide an adapter layer between kubevirt and
the virt-chroot binary backend implementation.

The creation of the tap device is currently using virt-chroot, due to
SELinux labeling requirements.

It is aimed to replace `CreateTapDevice` from `driver/common.go`.
The original is due to be removed once the setup flow stops using it.

Signed-off-by: Edward Haas <edwardh@redhat.com>