network: Introduce a new declarative API to configure the pod network
A new declarative API, that follows the nmstate specification [1].
The basic operations are `Read`, to view the existing configuration,
and `Apply`, to change the configuration.

It includes extensions per the needs of kubevirt (e.g. Linux Stack
parameters).

Integration tests have been included to cover basic scenarios.
These tests are marked as pending, and should be enabled and ran
manually (they require privileged permissions as they change the network
stack).

[1] https://nmstate.io/devel/yaml_api.html

Signed-off-by: Edward Haas <edwardh@redhat.com>