network, setup: Use the nmstate API to setup the pod network
Introduce a new NetPod object which is responsible to interpret the VMI
spec, read the existing pod network status and configure the pod network
accordingly.

NetPod uses the nmstate API to read and compose a desired state. Then
proceeds to mutate the pod network.

The masquerade configuration which involves nftables is not handled
through the nmstate API for now.

The change also includes a small logic change from the original:
For bridge binding, a dummy link is created regardless of IPAM.
(originally the dummy link was created only when IPAM was set)
This change provides a consistent link scheme with no need to
condition its existence. In addition, in follow-up changes the plan
is to use the dummy link for more usages.

Signed-off-by: Edward Haas <edwardh@redhat.com>