virt-controller, export: Use informers for instance types
This change switches over the export controller to use informers for
instance types, preferences and controller revisions after support was
recently introduced to expand the definition of a VirtualMachine [1].
This is in-line with the other controllers looking up these resources.

This commit is mostly `go fmt` fallout with the limited tests also being
moved over to using the informers to store and then fetch these
resources. More extensive tests should be added in future changes.

[1] https://github.com/kubevirt/kubevirt/pull/8953

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>