make: Introduce a new target to run integration tests
Functionality which requires OS/kernel services can now be validated
in a dedicated target: `integ-test`.

The tests are running in a container (isolated) which is privileged.
Tests are expected to be skipped if not explicitly directed to run
when executed in this context.
i.e. each integration test should include instructions to be skipped if
the `run-integration-tests` argument flag is not set.

To run the tests, execute `make integ-test`.

Signed-off-by: Edward Haas <edwardh@redhat.com>