Add a dedicated container for VMIs with dedicated CPUs
This new container would have a X dedicated CPUs while
the compute container can have 0.Y shared CPUs. In further
commits, the vCPU threads would be attached into the new
container's cgroup.

In order to support that some changes are introduced,
one of them is that the GetMemoryOverhead() function
is not split to two for vCPUs and other overhead.

Signed-off-by: Itamar Holder <iholder@redhat.com>