Make sure the VSOCK system service loads the correct module
There are two triggers to load vsock modules:

1. A syscall socket(AF_VSOCK, ...)
2. Accessing the vsock device file

In my case, AF_VSOCK on socket calls is used for vmware vsock and for
virtio vsock. The OS decides to load the vmware vsock when AF_VSOCK
syscall happens, and only one of the virtio drivers can be loaded at the same time. Therefore, it causes a problem.

If we access the virtio vsock device file first, udev has a clear signal what we actually want, the virtio vsock driver.

Signed-off-by: Zhuchen Wang <zcwang@google.com>