virt-api: Reject Instance types using GuestMappingPassthrough
This feature actually requires the use of dedictatedCPUPlacement that is
itself blocked by a lack of support for resource requests in instance
types. As we conflict when resource requests are present at the moment
there is no way of working around this limitation in a VirtualMachine
that uses instance types. As such for now until this support lands we
will reject instance type requests using GuestMappingPassthrough.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>