feat(rbac): Allow GET on VMI/portforward to kubevirt.io:[admin|edit]
Regular users with roles kubevirt.io:[admin|edit] need permission for
the GET method/verb on the virtualmachineinstances/portforward endpoint
to e.g. use portforwarding or virtctl ssh.

Conflicts:
  pkg/virt-operator/resource/generate/rbac/cluster.go
  tests/access_test.go

NOTE: This commit was changed from a fix to a feat because the initial
commit adding the permission with the wrong method/verb is not in this
branch.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>