virt-api: Refactor and share VM and VMI defaulting code between webhooks
This change moves the code responsible for defaulting the
VirtualMachineInstance and VirtualMachineInstanceSpec into a shared
location ahead of use by the VirtualMachine validation webhook. As a
result the code is also refactored to operate on the shared
VirtualMachineInstanceSpec. Helpers are also provided ensuring the
`defaulter-gen` generated code for each type is also called.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>